+++ server/router/api/v2/test/profile.js
@@ -1,6 +1,6 @@
 import express from "express";
-import {APIError} from "./routeBuilder/error/APIError.js";
-import {ErrorReason} from "./routeBuilder/error/ErrorReason.js";
+import {APIError} from "./routeBuilder/core/error/APIError.js";
+import {ErrorReason} from "./routeBuilder/core/error/ErrorReason.js";
 import {RouteBuilder} from "./routeBuilder/RouteBuilder.js";
 import {Method} from "./routeBuilder/core/enums/Method.js";
 import ProfileService from "../../../../service/Profile.js";
